+++ varframe/__init__.py
@@ -89,4 +89,4 @@ __all__ = [
     "ModelList",
 ]
 
-__version__ = "1.3.1"
+__version__ = "1.4.0"